% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_annex.R
\name{getvarnames}
\alias{getvarnames}
\title{getvarnames}
\usage{
getvarnames(formula, data = NULL)
}
\arguments{
\item{formula}{the formula from which to extract variable names}

\item{data}{data frame with variables from the formula}
}
\value{
'getvarnames' returns a list with 'varnames' (referring to all variable names), 'xvar' (referring to the predictors), and 'yvar' (referring to the outcome)
}
\description{
'getvarnames' is used to get all variable names from a regression model
}
