% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaChoice.R
\name{lambdaChoice}
\alias{lambdaChoice}
\title{lambdaChoice}
\usage{
lambdaChoice(
  treatment,
  mediators,
  outcome,
  lambdamax = 500,
  N1 = 10,
  selectedMin = 0.5 * length(outcome),
  selectedMax = length(outcome),
  L0 = 1,
  eta = 2,
  tau = 1,
  epsilon = 0.001
)
}
\arguments{
\item{treatment}{qsdfgh}

\item{mediators}{ertyu}

\item{outcome}{poiu}

\item{lambdamax}{mlkjh}

\item{N1}{kjhg}

\item{selectedMin}{mlkjh}

\item{selectedMax}{zerf}

\item{L0}{zert}

\item{eta}{dh}

\item{tau}{sdfghj}

\item{epsilon}{cdfg}
}
\value{
lambda value
}
\description{
"lambdaChoice" is used to .
}
