% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahi2.R
\name{mahi2}
\alias{mahi2}
\title{mahi2}
\usage{
mahi2(
  data,
  name.exposure,
  name.outcome,
  name.mediators,
  lambda,
  Nboot = 100,
  L0 = 0.1,
  eta = 2,
  tau = 1/P,
  epsilon = 0.001,
  Kmax = NULL,
  p.adjust.method = "bonferroni",
  pvalseuil = 0.05
)
}
\arguments{
\item{data}{a data.frame with the exposure, the outcome and mediators.}

\item{name.exposure}{a character string indicating the name of the exposure in data.}

\item{name.outcome}{a character string indicating the name of the outcome in data.}

\item{name.mediators}{a character string indicating the name of mediators in data.}

\item{lambda}{vector or a single value of lambda use for group-lasso.}

\item{Nboot}{number of bootstrap for stability selection of mediators. Default is 100.}

\item{L0}{dfghj}

\item{eta}{dfghj}

\item{tau}{dfghj}

\item{epsilon}{cvbn}

\item{Kmax}{maximum of mediators keept after the ranking obtained with stability selection.}

\item{p.adjust.method}{a character string indicating the name of the method for the correction for the multiple test. See help with p.adjust.methods.}

\item{pvalseuil}{the p-value for the multiple test}
}
\value{
mahi returns an object of class "mahi", a list that contains the components listed below.
}
\description{
'mahi2' is used to select mediators trough a high number of mediators and estimate various quantities for causal mediation analysis using "multimediate", with the selected mediators.
}
