% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahi3.R
\name{mahi3}
\alias{mahi3}
\title{mahi3}
\usage{
mahi3(
  data,
  name.exposure,
  name.outcome,
  name.mediators,
  lambda,
  Nboot = 100,
  L0 = 0.1,
  eta = 2,
  tau = 1/P,
  epsilon = 0.001,
  Kmax = NULL,
  bin = FALSE,
  step2 = FALSE,
  p.adjust.method = "bonferroni",
  pvalseuil = 0.05
)
}
\arguments{
\item{data}{a data.frame with the exposure, the outcome and mediators.}

\item{name.exposure}{a character string indicating the name of the exposure in data.}

\item{name.outcome}{a character string indicating the name of the outcome in data.}

\item{name.mediators}{a character string indicating the name of mediators in data.}

\item{lambda}{vector or a single value of lambda use for group-lasso.}

\item{Nboot}{number of bootstrap for stability selection of mediators. Default is 100.}

\item{L0}{dfghj}

\item{eta}{dfghj}

\item{tau}{dfghj}

\item{epsilon}{cvbn}

\item{Kmax}{maximum of mediators keept after the ranking obtained with stability selection.}

\item{bin}{a logical value indicating if the outcome is binary. if 'TRUE' a probit regression will be use is the second step. Default is 'FALSE'.}

\item{step2}{a logical value indicating if the a multiple test using the p-value calculated with multimediate (a multiple mediation analysis) have to be made. Default is 'FALSE'.}

\item{p.adjust.method}{a character string indicating the name of the method for the correction for the multiple test. See help with p.adjust.methods.}

\item{pvalseuil}{the p-value for the multiple test}
}
\value{
mahi returns an object of class "mahi", a list that contains the components listed below.
}
\description{
'mahi3' is used to select mediators trough a high number of mediators and estimate various quantities for causal mediation analysis using "multimediate", with the selected mediators.
}
