% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimediate.R
\name{multimediate}
\alias{multimediate}
\title{multimediate}
\usage{
multimediate(
  lmodel.m,
  correlated = FALSE,
  model.y,
  treat,
  treat.value = 1,
  control.value = 0,
  J = 1000,
  conf.level = 0.95
)
}
\arguments{
\item{lmodel.m}{list of fitted models object for mediators. Can be of class 'lm', 'polr','glm'.}

\item{correlated}{a logical value. if 'FALSE' a identity matrix is used for the matrix of correlation of mediators; if 'TRUE' matrix of correlation is estimated. Default is 'TRUE'.}

\item{model.y}{a fitted model object for the outcome. Can be of class 'lm', 'polr','glm'.}

\item{treat}{a character string indicating the name of the treatment variable used in the models. The treatment can be either binary (integer or a two-valued factor) or continuous (numeric).}

\item{treat.value}{value of the treatment variable used as the treatment condition. Default is 1.}

\item{control.value}{value of the treatment variable used as the control condition. Default is 0.}

\item{J}{number of Monte Carlo draws for quasi-Bayesian approximation.}

\item{conf.level}{level of the returned two-sided confidence intervals. Default is to return the 2.5 and 97.5 percentiles of the simulated quantities.}
}
\value{
multimediate returns an object of class "mm", a list that contains the components listed below.
The function summary (i.e., summary.mm) can be used to obtain a table of the results.
}
\description{
'multimediate' is used to estimate various quantities for causal mediation analysis, including average causal mediation effects (indirect effect), average direct effects, proportions mediated, and total effect, in presence of multiple mediators uncausally related. With a binary outcome, 'multimediate' also estimate average causal mediation effects on OR scale and logOR scale.
}
