% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bm.R
\name{summary.bm}
\alias{summary.bm}
\title{summary.bm}
\usage{
\method{summary}{bm}(object, opt = "navg", logit = "all", ...)
}
\arguments{
\item{object}{element of the class "blockmediate".}

\item{opt}{a character string indicating the details of the analysis "navg" for the average causal effects for t=0,1 and "avg"  for the average causal effects.}

\item{logit}{a character string indicating, when the outcome is binary, the scale of the average causal effects. "effects" for average causal effects, " OR" average causal effects on OR scale, "logOR" average causal effects on logOR scale and "all" for all scale.}

\item{...}{additional arguments affecting the summary produced}
}
\value{
table summarizing the causal analysis
}
\description{
"summary.bm" is used to display the results of the mediation analyzes done with "blockmediate".
}
