\name{data1}
\alias{data1}
\docType{data}
\title{
Toy dataset 1
}

\description{A toy dataset for causal mediation analysis with 3 uncausally related mediators.
}

\usage{data("data1")}
\format{
  A data frame with 1000 observations on the following 8 variables.
  \describe{
    \item{\code{Treatment}}{a factor with levels \code{0} \code{1}}
    \item{\code{C1}}{a factor with levels \code{0} \code{1}}
    \item{\code{C2}}{a factor with levels \code{0} \code{1}}
    \item{\code{C3}}{a factor with levels \code{0} \code{1}}
    \item{\code{M1}}{a numeric vector}
    \item{\code{M2}}{a numeric vector}
    \item{\code{M3}}{a numeric vector}
    \item{\code{Outcome}}{a numeric vector}
  }
}


\keyword{datasets}
