\name{data2}
\alias{data2}
\docType{data}
\title{
Toy dataset 2
}

\description{A toy dataset for causal mediation analysis with 3 uncausally related mediators.
}
\usage{data("data2")}
\format{
  A data frame with 1000 observations on the following 8 variables.
  \describe{
    \item{\code{Treatment}}{a factor with levels \code{0} \code{1}}
    \item{\code{C1}}{a factor with levels \code{0} \code{1}}
    \item{\code{C2}}{a factor with levels \code{0} \code{1}}
    \item{\code{C3}}{a factor with levels \code{0} \code{1}}
    \item{\code{M1}}{a numeric vector}
    \item{\code{M2}}{a numeric vector}
    \item{\code{M3}}{a numeric vector}
    \item{\code{Outcome}}{a factor with levels \code{FALSE} \code{TRUE}}
  }
}

\keyword{datasets}
