\name{data4}
\alias{data4}
\docType{data}
\title{
Toy dataset 4
}
\description{A toy dataset for causal mediation analysis with 3 uncausally related mediators.
}
\usage{data("data4")}
\format{
  A data frame with 1000 observations on the following 8 variables.
  \describe{
    \item{\code{Treatment}}{a numeric vector}
    \item{\code{C1}}{a factor with levels \code{0} \code{1}}
    \item{\code{C2}}{a factor with levels \code{0} \code{1}}
    \item{\code{C3}}{a factor with levels \code{0} \code{1}}
    \item{\code{M1}}{a numeric vector}
    \item{\code{M2}}{a factor with levels \code{0} \code{1}}
    \item{\code{M3}}{a factor with levels \code{0} \code{1} \code{2} \code{3}}
    \item{\code{Outcome}}{a numeric vector}
  }
}

\keyword{datasets}
