\name{data5}
\alias{data5}
\docType{data}
\title{
Toy dataset 5
}
\description{A toy dataset for causal mediation analysis with 3 uncausally related mediators
and survival outcome.
}
\usage{data("data5")}
\format{
  A data frame with 3000 observations on the following 8 variables.
  \describe{
    \item{\code{Exposure}}{a numeric vector}
    \item{\code{M1}}{a numeric vector}
    \item{\code{M2}}{a numeric vector}
    \item{\code{M3}}{a numeric vector}
    \item{\code{event}}{event}
    \item{\code{surv_time}}{time before event}
  }
}

\keyword{datasets}
