% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaChoice.R
\name{lambdaChoice}
\alias{lambdaChoice}
\title{lambdaChoice}
\usage{
lambdaChoice(
  data,
  name.exposure,
  name.outcome,
  name.mediators,
  lambdamax = 500,
  N1 = 10,
  selectedMin = 0.5 * length(outcome),
  selectedMax = length(outcome),
  L0 = 1,
  eta = 2,
  tau = 1,
  epsilon = 0.001
)
}
\arguments{
\item{data}{a data.frame with the exposure, the outcome and mediators.}

\item{name.exposure}{a character string indicating the name of the exposure in data.}

\item{name.outcome}{a character string indicating the name of the outcome in data.}

\item{name.mediators}{a character string indicating the name of mediators in data.}

\item{lambdamax}{mlkjh}

\item{N1}{kjhg}

\item{selectedMin}{mlkjh}

\item{selectedMax}{zerf}

\item{L0}{zert}

\item{eta}{dh}

\item{tau}{sdfghj}

\item{epsilon}{cdfg}
}
\value{
lambda value
}
\description{
"lambdaChoice" is used to .
}
